add google maps with pin markers to trailScreen, fix edge and pin order, and trying to make bookmarks work

+++ app/src/main/java/com/example/braguia/model/Pin.kt
@@ -5,25 +5,25 @@ import com.google.gson.annotations.SerializedName
 data class Pin(
 
     @SerializedName("id")
-    val id:Long,
+    val id: Long,
 
     @SerializedName("pin_name")
-    val pinName:String,
+    val pinName: String,
 
     @SerializedName("pin_desc")
-    val pinDesc:String,
+    val pinDesc: String,
 
     @SerializedName("pin_lat")
-    val pinLat:Float,
+    val pinLat: Double,
 
     @SerializedName("pin_lng")
-    val pinLng:Float,
+    val pinLng: Double,
 
     @SerializedName("pin_alt")
-    val pinAlt:Float,
+    val pinAlt: Double,
 
     @SerializedName("rel_pin")
-    val relPin:List<RelPin>,
+    val relPin: List<RelPin>,
 
     @SerializedName("media")
     val media: List<Media>
